ref #934 -- moved to Capitalizaed Directories
